Fix: character class (`[]`) cannot match an anchor & use sed -e
A character class (characters in a `[]`) always represents exact one
character, so no anchors can be put in it.
While ^ in the beginning of the character class performs a negation, other places escape it as a literal '^' rather than the beginning of a line.
If really needed, `(^|a)b` matches a or start of a line before b.
See
http://stackoverflow.com/questions/13199118/regex-line-start-in-character-set